Updates for the HC device tree (#5)
* Add the config for RTC

* Updates for the HC device tree

Some updates for the HC device tree source:

- standardise on AM33XX_IOPAD instead of AM33X_PADCONF as per the
  output from the TI SysConfig tool.
- standardise all pinmux comments to show the mux signal names and
  the CPU pin numbers.
- Remove entry for non-existent LED 'usr_7_led'.
- Remove RS485 parameters for UART2.
- Copy some missing entries across from the ALMP DTS.
- Update the signals based on the new HC schematic.
- Add provisional entries for the new I2C devices.
- Add provisional entry for the WiFi chipset.
- Add default settings for various enable/reset signals.

Signed-off-by: Mark Corbin <mark.corbin@moixa.com>

* The cd-gpios causes the kernel to not mount the SD card - removing for now

* Updates for the EEPROM and MMC DTS entries

Add additional information for the EEPROM devices. Change the
SD card 'card detect' entry to reflect the inverted signal. Add
the 'non-removable' attribute to the eMMC device. Update the MMC3
details for the BRCM WiFi module.

Signed-off-by: Mark Corbin <mark.corbin@moixa.com>

* Fix the enable of the USB +5V switch to the BG95

* Use GPIO defines (#6)

* Use gpio defines

* Fix USB1

* Modem pullup

* Correct the WiFi device power settings

Removed the unused wlan regulator entry. Set the WiFi MMC supply to
3.3v. Added some additional MMC attributes for the WiFi. Fixed the
comment tab spacing.

Signed-off-by: Mark Corbin <mark.corbin@moixa.com>

* Add default settings for USB power enable signals

Add default entries for USB0 and USB1 to set the power enable
signals so that power is provided to peripherals in host mode.

Signed-off-by: Mark Corbin <mark.corbin@moixa.com>

* Remove USB enable signals from DTS and update pull up/down setings

Signed-off-by: Mark Corbin <mark.corbin@moixa.com>

* Change compatibility for MAC EEPROM in DTS

The 24AA025E48T device isn't compatible with the Atmel 24mac02 so
we'll treat it as an ordinary 24C02 EEPROM.

Signed-off-by: Mark Corbin <mark.corbin@moixa.com>

* Add support for the AM335x SOC temperature sensor

Add support for the internal AM335x bandgap temperature sensor. Note
that this sensor has been widely reported as inaccurate and is not
recommended by TI - so use with caution! The temperature can be read
via the /sys/class/hwmon/hwmon0/device/temp1_input file.

Signed-off-by: Mark Corbin <mark.corbin@moixa.com>

Co-authored-by: Neil Wilson <neil.wilson@moixa.com>
Co-authored-by: Cesar Pecharroman Arribas <cesarpa@outlook.com>