Kernel option cleanup (#25)
* Drop kernel ATAG support

Kernel ATAGs have been superseded by the device tree.

Signed-off-by: Mark Corbin <mcorbin@lunarenergy.com>

* Drop UEFI support

Signed-off-by: Mark Corbin <mcorbin@lunarenergy.com>

* Drop PPS and PTP precision time support

Drop PPS and PTP support as we do not have any GPS receivers
or high precision clocks connected to the system.

Signed-off-by: Mark Corbin <mcorbin@lunarenergy.com>

* Remove support for unused filesystems

Remove support for unused filesystems. Change VFAT support to a
module. Leave NFS support compiled-in to the kernel so that we
have the option to remotely mount the rootfs if necessary.

Signed-off-by: Mark Corbin <mcorbin@lunarenergy.com>

* Drop PCI support

Drop PCI support as we don't have a PCI bus.

Signed-off-by: Mark Corbin <mcorbin@lunarenergy.com>

* Cleanup of USB drivers

Signed-off-by: Mark Corbin <mcorbin@lunarenergy.com>

* Add dummy USB regulator to the DTS file

Signed-off-by: Mark Corbin <mcorbin@lunarenergy.com>

* Update am335x-bandgap driver to use new kernel API

Signed-off-by: Mark Corbin <mcorbin@lunarenergy.com>

* Fix compiler version in defconfig file

Signed-off-by: Mark Corbin <mcorbin@lunarenergy.com>

Signed-off-by: Mark Corbin <mcorbin@lunarenergy.com>